refactor portfolio
