+++ packages/contracts-algebra-integral/src/portfolio/index.ts
@@ -1,1 +1,4 @@
-export * from "./getPortfolio.js";
+export * from "./balancePortfolio.js";
+export * from "./getBalancePortfolioAmounts.js";
+export * from "./getBalancePortfolioTrades.js";
+export * from "./getPortfolioHoldings.js";
