+++ packages/contracts-algebra-integral/src/swaprouter/decode.ts
@@ -1,8 +1,9 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 import { decodeFunctionData, Hex } from "viem";
 import { SwapRouter } from "../artifacts/SwapRouter.js";
 
-const USDC_ETH =
-    "0xac9650d8000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000016000000000000000000000000000000000000000000000000000000000000000e4bc651188000000000000000000000000d988097fb8612cc24eec14542bc03424c656005f0000000000000000000000004200000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000191c04de87400000000000000000000000000000000000000000000000000000000000f42400000000000000000000000000000000000000000000000000001771fe63519e4000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004469bc35b20000000000000000000000000000000000000000000000000001771fe63519e4000000000000000000000000fe732ca712c695ee14a8a015e65997dd9189c31b00000000000000000000000000000000000000000000000000000000";
+// const USDC_ETH =
+// "0xac9650d8000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000040000000000000000000000000000000000000000000000000000000000000016000000000000000000000000000000000000000000000000000000000000000e4bc651188000000000000000000000000d988097fb8612cc24eec14542bc03424c656005f0000000000000000000000004200000000000000000000000000000000000006000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000191c04de87400000000000000000000000000000000000000000000000000000000000f42400000000000000000000000000000000000000000000000000001771fe63519e4000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000004469bc35b20000000000000000000000000000000000000000000000000001771fe63519e4000000000000000000000000fe732ca712c695ee14a8a015e65997dd9189c31b00000000000000000000000000000000000000000000000000000000";
 const ETH_USDC =
     "0xac9650d800000000000000000000000000000000000000000000000000000000000000200000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000002000000000000000000000000000000000000000000000000000000000000000e4bc6511880000000000000000000000004200000000000000000000000000000000000006000000000000000000000000d988097fb8612cc24eec14542bc03424c656005f000000000000000000000000fe732ca712c695ee14a8a015e65997dd9189c31b00000000000000000000000000000000000000000000000000000191adea7b2e0000000000000000000000000000000000000000000000000011c37937e080000000000000000000000000000000000000000000000000000000000000bad171000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000";
